fix: place link actions beside address bar
