+++ apps/canvas-workspace/src/renderer/src/components/LinkDrawer/index.css
@@ -42,6 +42,63 @@
   color: var(--text);
 }
 
+.link-drawer__actions {
+  display: inline-flex;
+  align-items: center;
+  gap: 5px;
+  flex: 0 0 auto;
+}
+
+.link-drawer__action {
+  appearance: none;
+  width: 28px;
+  height: 28px;
+  padding: 0;
+  border: 1px solid rgba(55, 53, 47, 0.12);
+  border-radius: var(--radius-xs);
+  background: #fff;
+  color: #73716c;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+  transition:
+    background 0.14s ease,
+    border-color 0.14s ease,
+    color 0.14s ease,
+    transform 0.2s ease;
+}
+
+.link-drawer__action:hover:not(:disabled) {
+  background: rgba(55, 53, 47, 0.06);
+  border-color: rgba(55, 53, 47, 0.2);
+  color: #37352f;
+  transform: translateY(-1px);
+}
+
+.link-drawer__action:focus-visible {
+  outline: 2px solid rgba(35, 131, 226, 0.38);
+  outline-offset: 2px;
+}
+
+.link-drawer__action--primary {
+  border-color: #2383e2;
+  background: #2383e2;
+  color: #fff;
+  box-shadow: 0 4px 10px rgba(35, 131, 226, 0.18);
+}
+
+.link-drawer__action--primary:hover:not(:disabled) {
+  border-color: #1f76cc;
+  background: #1f76cc;
+}
+
+.link-drawer__action:disabled {
+  opacity: 0.55;
+  cursor: not-allowed;
+  box-shadow: none;
+}
+
 /* ── Webview host ───────────────────────────────────────────────────── */
 
 .link-drawer__webview-host {
